pyproject.toml
[project]
name = "langgraph-invoice-ai"
version = "0.1.0"
description = "Agentic multi-agent invoice reconciliation system using LangGraph, OCR, and LLM reasoning."
readme = "README.md"
requires-python = ">=3.10"

authors = [
  { name="Amgoth Sunil", email="amgothsunil422@email.com" }
]

dependencies = [
    "easyocr>=1.7.2",
    "fastapi>=0.128.0",
    "fuzzywuzzy>=0.18.0",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "loguru>=0.7.3",
    "numpy>=2.2.6",
    "openai>=2.16.0",
    "opencv-python>=4.13.0.90",
    "paddleocr>=3.4.0",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-levenshtein>=0.27.3",
    "python-multipart>=0.0.22",
    "rapidfuzz>=3.14.3",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.53.1",
    "tenacity>=9.1.2",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.4.0",
    "ruff>=0.5.0",
    "pytest>=8.2.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

# -------------------
# TOOLING CONFIG
# -------------------

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
